# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Toyo Kambocha HRMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List departments
      responses:
        "200":
          description: Departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    put:
      operationId: updateDepartment
      tags: [departments]
      summary: Update department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Mark attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [attendance]
      summary: List leave requests
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: applyLeave
      tags: [attendance]
      summary: Apply for leave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyLeaveRequest"
      responses:
        "201":
          description: Leave applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}/status:
    put:
      operationId: updateLeaveStatus
      tags: [attendance]
      summary: Approve or reject leave
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeaveStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        headId:
          type: integer
          nullable: true
        headName:
          type: string
          nullable: true
        employeeCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, employeeCount, createdAt]
    CreateDepartmentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        headId:
          type: integer
          nullable: true
      required: [name]
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        date:
          type: string
        checkIn:
          type: string
          nullable: true
        checkOut:
          type: string
          nullable: true
        status:
          type: string
          enum: [present, absent, late, half_day, on_leave]
        workHours:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
      required: [id, employeeId, employeeName, date, status]
...
    CreateAttendanceRequest:
      type: object
      properties:
        employeeId:
          type: integer
        date:
          type: string
        checkIn:
          type: string
          nullable: true
        checkOut:
          type: string
          nullable: true
        status:
          type: string
          enum: [present, absent, late, half_day, on_leave]
        notes:
          type: string
          nullable: true
      required: [employeeId, date, status]
    LeaveRequest:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        leaveType:
          type: string
          enum: [annual, sick, casual, maternity, paternity, unpaid]
        startDate:
          type: string
        endDate:
          type: string
        days:
          type: number
        reason:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        approvedBy:
          type: integer
          nullable: true
        approvedByName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, employeeId, employeeName, leaveType, startDate, endDate, days, reason, status, createdAt]
    ApplyLeaveRequest:
      type: object
      properties:
        employeeId:
          type: integer
        leaveType:
          type: string
          enum: [annual, sick, casual, maternity, paternity, unpaid]
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
      required: [employeeId, leaveType, startDate, endDate, reason]
    UpdateLeaveStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
        approvedBy:
          type: integer
      required: [status, approvedBy]
...