# lib/api-zod/src/generated/types/createAttendanceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Toyo Kambocha HRMS API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAttendanceRequestStatus } from "./createAttendanceRequestStatus";

export interface CreateAttendanceRequest {
  employeeId: number;
  date: string;
  checkIn?: string | null;
  checkOut?: string | null;
  status: CreateAttendanceRequestStatus;
  notes?: string | null;
}